Extract some more logic.
